fix(voice): extract assets panel and silence lint for JOE-1109
